Separate PathClassManager from QuPathGUI
